+++ plotlykt-core/api/plotlykt-core.api
@@ -24,9 +24,8 @@ public final class space/kscience/plotly/FileExportKt {
 }
 
 public final class space/kscience/plotly/HtmlKt {
+	public static final fun getCdnPlotlyHeader ()Lspace/kscience/plotly/PlotlyHtmlFragment;
 	public static final fun getMathJaxHeader ()Lspace/kscience/plotly/PlotlyHtmlFragment;
-	public static final fun localCssHeader (Ljava/nio/file/Path;Ljava/nio/file/Path;Ljava/lang/String;)Lspace/kscience/plotly/PlotlyHtmlFragment;
-	public static final fun localScriptHeader (Ljava/nio/file/Path;Ljava/nio/file/Path;Ljava/lang/String;)Lspace/kscience/plotly/PlotlyHtmlFragment;
 	public static final fun plus (Lspace/kscience/plotly/PlotlyHtmlFragment;Lspace/kscience/plotly/PlotlyHtmlFragment;)Lspace/kscience/plotly/PlotlyHtmlFragment;
 	public static final fun toHTML (Lspace/kscience/plotly/Plot;[Lspace/kscience/plotly/PlotlyHtmlFragment;Lspace/kscience/plotly/PlotlyConfig;)Ljava/lang/String;
 	public static synthetic fun toHTML$default (Lspace/kscience/plotly/Plot;[Lspace/kscience/plotly/PlotlyHtmlFragment;Lspace/kscience/plotly/PlotlyConfig;ILjava/lang/Object;)Ljava/lang/String;
@@ -180,7 +179,8 @@ public final class space/kscience/plotly/PlotlyFragment {
 }
 
 public final class space/kscience/plotly/PlotlyHeadersKt {
-	public static final fun getCdnPlotlyHeader ()Lspace/kscience/plotly/PlotlyHtmlFragment;
+	public static final fun localCssHeader (Ljava/nio/file/Path;Ljava/nio/file/Path;Ljava/lang/String;)Lspace/kscience/plotly/PlotlyHtmlFragment;
+	public static final fun localScriptHeader (Ljava/nio/file/Path;Ljava/nio/file/Path;Ljava/lang/String;)Lspace/kscience/plotly/PlotlyHtmlFragment;
 }
 
 public final class space/kscience/plotly/PlotlyHtmlFragment {
